Support for pull-down resistor on pin
Mirrors the activation of the pull-up resistor in that writing LOW
to a pin will enable the built-in pull-down resistor. This does
no checking to see if the pin actually has the required resistor.

The limitation of this API is that once set, there's no way to
set the pin back to the no-resistor state.